refactor: move error handling to common/
